homework-8.1 - исправления по комментариям проверяюшего (храние коментарии в отдельной коллекции,добавляем слой интерфейс в service)
