+++ home-8/src/main/java/ru/otus/home8/domain/Book.java
@@ -22,7 +22,6 @@ public class Book {
     @DBRef
     private Genre genre;
 
-    private List<BookComment> comments = new ArrayList<>();
 }
