+++ home-8/src/main/java/ru/otus/home8/mongock/DataBaseChangeLog.java
@@ -44,8 +44,6 @@ public class DataBaseChangeLog {
                 .append("bookId", bookId.toString());
         commentCollection.insertOne(commentDoc);
 
-        bookCollection.updateOne(eq("_id", bookId), push("comments", commentDoc));
-
     }
